requirements.txt
# AI agent
pandas==2.3.0
openai==1.90.0
anthropic==0.54.0
langchain==0.3.26
langchain-openai==0.3.24
langchain-anthropic==0.3.15
langchain-community==0.3.26
langchain-experimental==0.3.4
langchain-core==0.3.68
langgraph==0.4.8
langgraph-prebuilt==0.2.2
paho-mqtt==2.1.0
together==1.5.17
notebook==7.4.3

# RTD / chart generation
beautifulsoup4==4.13.3
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
fonttools==4.56.0
google==3.0.0
google-api-core==2.24.2
google-auth==2.38.0
google-cloud==0.34.0
google-cloud-speech==2.31.0
google-cloud-texttospeech==2.25.0
googleapis-common-protos==1.69.1
grpcio==1.71.0
grpcio-status==1.71.0
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.1
networkx==3.4.2
numpy==2.2.3
packaging==24.2
pillow==11.1.0
proto-plus==1.26.1
PyAudio==0.2.14
pygame==2.6.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
requests==2.32.3
rsa==4.9
six==1.17.0
soupsieve==2.6
typing_extensions==4.12.2
urllib3==2.3.0
